1.3 UserService

+++ src/main/java/com/soCompany/sunflower/repository/CommunityMemberRepository.java
@@ -3,5 +3,5 @@ package com.soCompany.sunflower.repository;
 import com.soCompany.sunflower.entity.CommunityMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommunityMemberRepository extends JpaRepository<CommunityMember, Long> {
+public interface CommunityMemberRepository extends JpaRepository<CommunityMember, Integer> {
 }
