+++ src/main/java/com/soCompany/sunflower/repository/CommunityRepository.java
@@ -3,5 +3,5 @@ package com.soCompany.sunflower.repository;
 import com.soCompany.sunflower.entity.Community;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommunityRepository extends JpaRepository<Community, Long> {
+public interface CommunityRepository extends JpaRepository<Community, Integer> {
 }
